+++ src/app/pages/Collectibles/components/VideoCollectible.tsx
@@ -3,7 +3,7 @@ import React, { FC, useCallback, useState } from 'react';
 import { emptyFn } from '@rnw-community/shared';
 import clsx from 'clsx';
 
-import { VideoPlayer } from './VideoPlayer/VideoPlayer';
+import { Player } from './VideoPlayer/Player';
 
 interface Props {
   uri: string;
@@ -19,9 +19,8 @@ export const VideoCollectible: FC<Props> = ({ uri, loader, className, onError = 
 
   return (
     <>
-      <VideoPlayer
+      <Player
         src={uri}
-        autoPlay
         loop
         onCanPlayThrough={handleLoaded}
         className={clsx(className, isLoading && 'hidden')}
