+++ src/app/pages/Collectibles/components/VideoPlayer/components/Progress.tsx
@@ -2,7 +2,6 @@ import React, { memo } from 'react';
 
 interface ProgressProps {
   videoDuration: number;
-  bufferProgress: number;
   currentProgress: number;
   seekProgress: number;
   seekTooltipPosition: string;
@@ -12,21 +11,11 @@ interface ProgressProps {
 }
 
 export const Progress = memo<ProgressProps>(
-  ({
-    bufferProgress,
-    currentProgress,
-    videoDuration,
-    seekProgress,
-    seekTooltipPosition,
-    seekTooltip,
-    onHover,
-    onSeek
-  }) => {
+  ({ currentProgress, videoDuration, seekProgress, seekTooltipPosition, seekTooltip, onHover, onSeek }) => {
     return (
       <div className="vp-progress">
         <div className="vp-progress__range">
           <div className="vp-progress__range--background" />
-          <div className="vp-progress__range--buffer" style={{ width: bufferProgress + '%' }} />
           <div className="vp-progress__range--current" style={{ width: currentProgress + '%' }}>
             <div className="vp-progress__range--current__thumb" />
           </div>
